requirements.txt
colorama>=0.4.6
requests>=2.31.0
beautifulsoup4>=4.12.2
googlesearch-python>=1.2.3
rich>=13.7.0